+++ functions.php
@@ -91,7 +91,6 @@ function twentyseventeen_setup() {
 endif;
 add_action( 'after_setup_theme', 'twentyseventeen_setup' );
 
-
 /**
  * Set the content width in pixels, based on the theme's design and stylesheet.
  *
@@ -103,7 +102,6 @@ function twentyseventeen_content_width() {
 	$GLOBALS['content_width'] = apply_filters( 'twentyseventeen_content_width', 700 );
 }
 add_action( 'after_setup_theme', 'twentyseventeen_content_width', 0 );
-
 
 /**
  * Register custom fonts
@@ -172,7 +170,6 @@ function twentyseventeen_widgets_init() {
 }
 add_action( 'widgets_init', 'twentyseventeen_widgets_init' );
 
-
 /**
  * Output Custom Logo.
  */
@@ -181,7 +178,6 @@ function twentyseventeen_the_custom_logo() {
 		the_custom_logo();
 	}
 }
-
 
 if ( ! function_exists( 'twentyseventeen_excerpt_continue_reading' ) ) {
 /**
@@ -205,6 +201,33 @@ function twentyseventeen_javascript_detection() {
 	echo "<script>(function(html){html.className = html.className.replace(/\bno-js\b/,'js')})(document.documentElement);</script>\n";
 }
 add_action( 'wp_head', 'twentyseventeen_javascript_detection', 0 );
+
+/**
+ * Add a pingback url auto-discovery header for singularly identifiable articles.
+ */
+function twentyseventeen_pingback_header() {
+	if ( is_singular() && pings_open() ) {
+		printf( '<link rel="pingback" href="%s">' . "\n", get_bloginfo( 'pingback_url' ) );
+	}
+}
+add_action( 'wp_head', 'twentyseventeen_pingback_header' );
+
+/**
+ * Display custom color CSS.
+ */
+function twentyseventeen_colors_css_wrap() {
+	if ( 'custom' !== get_theme_mod( 'colorscheme' ) && ! is_customize_preview() ) {
+		return;
+	}
+
+	require_once( 'color-patterns.php' );
+	$hue = absint( get_theme_mod( 'colorscheme_hue', 250 ) );
+?>
+	<style type="text/css" id="custom-theme-colors" <?php if ( is_customize_preview() ) { echo 'data-hue="' . $hue . '"'; } ?>>
+		<?php echo twentyseventeen_custom_colors_css(); ?>
+	</style>
+<?php }
+add_action( 'wp_head', 'twentyseventeen_colors_css_wrap' );
 
 /**
  * Enqueue scripts and styles.
